Inventory_management/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
SQLAlchemy==2.0.49
psycopg2-binary==2.9.12
pydantic==2.13.3
python-dotenv==1.2.2
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.18.4
email-validator==2.3.0